requirement.txt
boto==2.38.0
bz2file==0.98
cffi==1.3.1
cycler==0.9.0
Cython==0.23.4
cytoolz==0.7.4
decorator==4.0.4
future==0.15.2
gensim==0.12.3
httpretty==0.8.6
matplotlib==1.5.0
natto-py==0.7.0
networkx==1.10
nltk==3.1
nose==1.3.7
numpy==1.10.1
oauthlib==1.0.3
pycparser==2.14
pyparsing==2.0.6
python-dateutil==2.4.2
python-twitter==2.3
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
scikit-learn==0.17
scipy==0.16.1
six==1.10.0
sklearn==0.0
smart-open==1.3.0
SQLAlchemy==1.0.9
Theano==0.7.0
